Iterative decoder Type is now np.uint8

+++ GERCE.py
@@ -25,7 +25,7 @@ def GERCE(M, Niter = 1):
     # cur_rank = 0
 
     for i in range(Niter):
-        Q = np.identity(n, dtype=np.int64)  # reset ECO mat
+        Q = np.identity(n, dtype=np.uint8)  # reset ECO mat
         if i == 0: # no permutation for first loop
             R = M
         else:
